Feat/pause resume

+++ Sources/mobile-offline-downloader-ios/Managers/OfflineDownloadsManager.swift
@@ -11,7 +11,11 @@ public struct OfflineDownloadsManagerEventObject {
     public var status: OfflineDownloaderStatus
     public var progress: Double
 }
-        
+
+public enum OfflineDownloadsQueueEvent {
+    case completed(success: Bool)
+}
+
 public class OfflineDownloadsManager {
     public static var shared: OfflineDownloadsManager = .init()
 
@@ -23,17 +27,32 @@ public class OfflineDownloadsManager {
 
     var entries: [OfflineDownloaderEntry] = []
     public var activeEntries: [OfflineDownloaderEntry] {
-        downloaders
-            .filter { $0.status == .active || $0.status == .initialized || $0.status == .preparing }
-            .map { $0.entry }
+        entries
+            .filter { $0.status == .active || $0.status == .preparing }
+            .sorted(by: { $0.updatedTimestamp > $1.updatedTimestamp })
     }
 
-    var completedEntries: [OfflineDownloaderEntry] {
-        []
+    public var completedEntries: [OfflineDownloaderEntry] {
+        entries
+            .filter { $0.status == .completed }
     }
 
-    var pausedEntries: [OfflineDownloaderEntry] {
-        []
+    public var waitingEntries: [OfflineDownloaderEntry] {
+        entries
+            .filter { $0.status == .initialized }
+            .sorted(by: { $0.updatedTimestamp > $1.updatedTimestamp })
+    }
+    
+    public var pausedEntries: [OfflineDownloaderEntry] {
+        entries
+            .filter { $0.status == .paused }
+            .sorted(by: { $0.updatedTimestamp > $1.updatedTimestamp })
+
+    }
+    
+    public var failedEntries: [OfflineDownloaderEntry] {
+        entries
+            .filter { $0.status == .failed }
     }
 
     var downloaders: [OfflineEntryDownloader] = []
@@ -45,14 +64,35 @@ public class OfflineDownloadsManager {
             .share()
             .eraseToAnyPublisher()
     }()
-    
+    private var sourceQueuePublisher: PassthroughSubject<OfflineDownloadsQueueEvent, Never> = .init()
+    lazy public var queuePublisher: AnyPublisher<OfflineDownloadsQueueEvent, Never> =  {
+        sourceQueuePublisher
+            .receive(on: DispatchQueue.main)
+            .share()
+            .eraseToAnyPublisher()
+    }()
+
     init() {
         updateFolder()
+        loadEntries()
     }
-    
+
+    private func loadEntries() {
+        OfflineStorageManager.shared.loadAll(of: OfflineDownloaderEntry.self) { result in
+            if case .success(let entries) = result {
+                self.entries = entries
+                for entry in self.activeEntries + self.waitingEntries {
+                    self.start(entry: entry)
+                }
+            } else {
+                // TODO: failed state
+            }
+        }
+    }
+
     private func updateFolder() {
         do {
-                // exclude from cloud backup
+            // exclude from cloud backup
             var cacheURL = config.rootPath.fileURL(isDirectory: true)
             let existingValues = try cacheURL.resourceValues(forKeys: [.isExcludedFromBackupKey])
             if existingValues.isExcludedFromBackup == false || existingValues.isExcludedFromBackup == nil {
@@ -69,39 +109,115 @@ public class OfflineDownloadsManager {
         self.config = config
     }
 
-    public func addAndStart(object: OfflineDownloadTypeProtocol, userInfo: String? = nil) throws {
+    public func start(object: OfflineDownloadTypeProtocol, userInfo: String? = nil) throws {
         let entry = try object.downloaderEntry()
         entry.userInfo = userInfo
-        guard getEntry(for: entry.dataModel.id, type: entry.dataModel.type) == nil else { return }
-        entries.append(entry)
+        if getQueuedEntry(for: entry) == nil {
+            entries.append(entry)
+            entry.saveToDB(completion: {_ in })
+        }
         start(entry: entry)
+    }
+    
+    private func sendStatusEvent(for entry: OfflineDownloaderEntry, progress: CGFloat = 0) {
+        guard let object = object(for: entry.dataModel) else { return }
+        let publisherObject = OfflineDownloadsManagerEventObject(
+            object:  object,
+            status: entry.status,
+            progress: progress
+        )
+        sourcePublisher.send(.statusChanged(object: publisherObject))
     }
 
     private func start(entry: OfflineDownloaderEntry) {
-        guard getEntry(for: entry.dataModel.id, type: entry.dataModel.type) != nil else { return }
-        if entry.status != .completed && activeEntries.count < config.limitOfConcurrentDownloads {
-            if let downloader = getDownloader(for: entry) {
+        guard let entry = getQueuedEntry(for: entry) else { return }
+        if entry.status != .completed {
+            if entry.status != .initialized {
+                entry.status = .initialized
+            }
+            
+            sendStatusEvent(for: entry)
+
+            if activeEntries.count < config.limitOfConcurrentDownloads {
+                let downloader = getDownloader(for: entry) ?? createDownloader(for: entry)
                 if downloader.status.canStart {
                     downloader.start()
                 }
-            } else {
-                let downloader = createDownloader(for: entry)
-                downloaders.append(downloader)
-                downloader.start()
             }
         }
     }
+    
+    private func startNext(latestStatus: OfflineDownloaderStatus) {
+        guard let entry = waitingEntries.first else {
+            if activeEntries.isEmpty {
+                if !failedEntries.filter({ !$0.errors.isEmpty }).isEmpty {
+
+                    sourceQueuePublisher.send(.completed(success: false))
+
+                    failedEntries.forEach {
+                        $0.errors = []
+                    }
+                } else if latestStatus == .completed {
+                    sourceQueuePublisher.send(.completed(success: true))
+                }
+            }
+            return
+        }
+        start(entry: entry)
+    }
+
+    public func pause(object: OfflineDownloadTypeProtocol) throws {
+        let entry = try object.downloaderEntry()
+        pause(entry: entry)
+    }
 
     public func pause(entry: OfflineDownloaderEntry) {
+        if let downloader = getDownloader(for: entry) {
+            downloader.pause()
+        } else if let entry = getQueuedEntry(for: entry) {
+            entry.status = .paused
+            sendStatusEvent(for: entry)
+        }
+    }
 
+    public func cancel(object: OfflineDownloadTypeProtocol) throws {
+        let entry = try object.downloaderEntry()
+        cancel(entry: entry)
     }
 
     public func cancel(entry: OfflineDownloaderEntry) {
+        stopLoading(entry: entry)
+        removeFromQueue(entry: entry)
+    }
 
+    public func deleteDownloadingEntries() throws {
+        try entries.filter {
+            switch $0.status {
+            case .initialized, .active, .preparing, .paused, .failed, .cancelled:
+                return true
+            default:
+                return false
+            }
+        }.forEach {
+            try delete(entry: $0)
+        }
     }
 
     public func resume(entry: OfflineDownloaderEntry) {
-        
+        start(entry: entry)
+    }
+    
+    public func resume(object: OfflineDownloadTypeProtocol) throws {
+        let entry = try object.downloaderEntry()
+        resume(entry: entry)
+    }
+    
+    func getQueuedEntry(for entry: OfflineDownloaderEntry) -> OfflineDownloaderEntry? {
+        getEntry(for: entry.dataModel)
+    }
+    
+    func getEntry(for dataModel: OfflineStorageDataModel) -> OfflineDownloaderEntry? {
+        getEntry(for: dataModel.id, type: dataModel.type)
     }
     
     func getEntry(for id: String, type: String) -> OfflineDownloaderEntry? {
@@ -128,35 +244,53 @@ public class OfflineDownloadsManager {
     }
 
     public func delete(entry: OfflineDownloaderEntry) throws {
-        if let index = entries.firstIndex(where: { $0.dataModel.id == entry.dataModel.id }) {
-            let managerEntry = entries[index]
-            if let downloader = getDownloader(for: managerEntry) {
-                downloader.cancel()
+        cancel(entry: entry)
+        try removeLocalFiles(for: entry)
+        removeFromStorage(entry: entry)
+        startNext(latestStatus: .removed)
+    }
 
-                if let index = downloaders.firstIndex(of: downloader) {
-                    downloaders.remove(at: index)
-                }
-            }
-            entries.remove(at: index)
-        }
-        let path = entry.rootPath(with: config.rootPath)
-        try FileManager.default.removeItem(atPath: path)
-
+    private func removeFromStorage(entry: OfflineDownloaderEntry) {
         OfflineStorageManager.shared.delete(entry) {[weak self] result in
-            guard let self = self, let object = self.object(for: entry.dataModel) else {
-                return
-            }
             if case .success = result {
-                let publisherObject = OfflineDownloadsManagerEventObject(
-                    object:  object,
-                    status: .removed,
-                    progress: 0
-                )
-                self.sourcePublisher.send(.statusChanged(object: publisherObject))
+                entry.status = .removed
+                self?.sendStatusEvent(for: entry)
             }
         }
     }
-    
+
+    private func removeFromQueue(entry: OfflineDownloaderEntry) {
+        // Search entry in queue
+        if let index = entries.firstIndex(where: { $0.dataModel.id == entry.dataModel.id }) {
+
+            // remove downloader
+            removeDownloader(for: entry)
+            
+            // remove entry
+            entries.remove(at: index)
+        }
+    }
+
+    private func stopLoading(entry: OfflineDownloaderEntry) {
+        guard let downloader = getDownloader(for: entry) else { return }
+        downloader.cancel()
+    }
+
+    private func removeDownloader(for entry: OfflineDownloaderEntry) {
+        guard let downloader = getDownloader(for: entry) else { return }
+        if let index = downloaders.firstIndex(of: downloader) {
+            downloaders.remove(at: index)
+        }
+    }
+
+    private func removeLocalFiles(for entry: OfflineDownloaderEntry) throws {
+        // clear entry directory
+        let path = entry.rootPath(with: config.rootPath)
+        if FileManager.default.fileExists(atPath: path) {
+            try FileManager.default.removeItem(atPath: path)
+        }
+    }
+
     public func savedEntry<T: OfflineDownloadTypeProtocol>(for object: T, completionHandler: @escaping(Result<OfflineDownloaderEntry, Error>) -> Void) {
         do {
             let dataModel = try object.toOfflineModel()
@@ -168,7 +302,7 @@ public class OfflineDownloadsManager {
             completionHandler(.failure(error))
         }
     }
-    
+
     public func savedValue(for entry: OfflineDownloaderEntry, pageIndex: Int ) -> OfflineDownloaderSavedValue {
         guard pageIndex < entry.parts.count && pageIndex >= 0 else { return .unknown }
         let part = entry.parts[pageIndex]
@@ -189,11 +323,16 @@ public class OfflineDownloadsManager {
     public func eventObject(for object: OfflineDownloadTypeProtocol, completionBlock: @escaping (Result<OfflineDownloadsManagerEventObject, Error>) -> Void) {
         do {
             let entry = try object.downloaderEntry()
-            if let entry = getEntry(for: entry.dataModel.id, type: entry.dataModel.type),
-                let downloader = getDownloader(for: entry){
-                let progress = downloader.progress.fractionCompleted
-                let eventObject = OfflineDownloadsManagerEventObject(object: object, status: downloader.status, progress: progress)
-                completionBlock(.success(eventObject))
+            if let entry = getQueuedEntry(for: entry) {
+                if let downloader = getDownloader(for: entry) {
+                    let progress = downloader.progress.fractionCompleted
+                    let eventObject = OfflineDownloadsManagerEventObject(object: object, status: downloader.status, progress: progress)
+                    completionBlock(.success(eventObject))
+                } else {
+                    let progress: Double = entry.status == .completed ? 1 : 0
+                    let eventObject = OfflineDownloadsManagerEventObject(object: object, status: entry.status, progress: progress)
+                    completionBlock(.success(eventObject))
+                }
                 return
             }
 
@@ -212,9 +351,17 @@ public class OfflineDownloadsManager {
         }
     }
     
+    public func canDownload(object: OfflineDownloadTypeProtocol) -> Bool {
+        if let entry = try? object.downloaderEntry() {
+            let type = type(of: object)
+            return type.canDownload(entry: entry)
+        }
+        return false
+    }
+    
     func createDownloader(for entry: OfflineDownloaderEntry) -> OfflineEntryDownloader {
         let downloader = OfflineEntryDownloader(entry: entry, config: config)
-        
+        downloaders.append(downloader)
         downloader.publisher(for: \.progress.fractionCompleted)
             .receive(on: DispatchQueue.main)
             .sink {[weak self, weak downloader] fractionCompleted in
@@ -227,23 +374,16 @@ public class OfflineDownloadsManager {
         downloader.statusPublisher
             .receive(on: DispatchQueue.main)
             .sink {[weak self, weak downloader] status in
-                guard let downloader = downloader, let object = self?.object(for: downloader.entry.dataModel) else { return }
-
-                let publisherObject = OfflineDownloadsManagerEventObject(object: object, status: status, progress: downloader.progress.fractionCompleted)
-                self?.sourcePublisher.send(.statusChanged(object: publisherObject))
-                switch status {
-                case .completed:
-                    print("publisherObject completed")
-                    OfflineStorageManager.shared.save(entry) { result in
-                        switch result {
-                        case .success:
-                            print("✅ Download finished and saved to Realm")
-                        case .failure(let error):
-                            print("⚠️ Download finished but Realm failed error = \(error)")
-                        }
-                    }
-                default:
-                    break
+                guard let downloader = downloader else { return }
+                
+                self?.sendStatusEvent(for: downloader.entry, progress: downloader.progress.fractionCompleted)
+                
+                if status == .completed || status == .failed || status == .cancelled || status == .paused {
+                    self?.removeDownloader(for: downloader.entry)
+                }
+                
+                if status == .completed || status == .paused || status == .failed || status == .cancelled {
+                    self?.startNext(latestStatus: status)
                 }
             }
             .store(in: &cancellables)
